Check equality on data and key attributes of im3d objects

Check equality on data and key attributes of im3d objects

# S3 method for im3d
all.equal(target, current, tolerance = 1e-06,
  attrsToCheck = c("BoundingBox"), attrsToCheckIfPresent = c("dim", "names",
  "dimnames", "x", "y", "z"), CheckSharedAttrsOnly = FALSE, ...)

Arguments

target

R object.

current

other R object, to be compared with target.

tolerance

numeric \(\ge\) 0. Differences smaller than tolerance are not reported. The default value is close to 1.5e-8.

attrsToCheck

Which attributes in im3d should always be checked

attrsToCheckIfPresent

Which attributes in im3d should be checked if present

CheckSharedAttrsOnly

Logical whether to check shared attributes only (default: FALSE)

...

additional arguments passed to all.equal

See also

all.equal