Convert an object to a nat hxsurf object

Convert an object to a nat hxsurf object

as.hxsurf(x, ...)

# S3 method for mesh3d
as.hxsurf(x, region = "Interior", col = NULL, ...)

Arguments

x
A surface object

...
Additional arguments passed to methods

region
The default name for the surface region

col
The surface colour (default value of NULL implies the colour specified in mesh3d object or grey when the mesh3d object has no colour.)

Value

A new surface object of class hxsurf (see read.hxsurf) for details.

Details

hxsurf objects are based on the format of Amira's surface objects (see read.hxsurf). They have the ability to include multiple distinct regions. However, at the moment the only method that we provide converts mesh3d objects, which can only include one region.

See also

as.mesh3d

Other hxsurf: as.mesh3d, materials, plot3d.hxsurf, read.hxsurf, subset.hxsurf, write.hxsurf

Examples

tet=tetrahedron3d(col='red')
teth=as.hxsurf(tet)

plot3d(teth)