Convert CMTK registration to homogeneous affine matrix with dof2mat

Convert CMTK registration to homogeneous affine matrix with dof2mat

cmtk.dof2mat(reg, Transpose = TRUE, version = FALSE)

Arguments

reg
Path to input registration file or 5x3 matrix of CMTK parameters.

Transpose
ouput matrix so that form on disk matches R's convention.

version
Whether to return CMTK version string

Value

4x4 transformation matrix

Details

Transpose is true by default since this results in the orientation of cmtk output files matching the orientation in R. Do not change this unless you're sure you know what you're doing!

See also

Other cmtk-commandline: cmtk.mat2dof

Other cmtk-geometry: affmat2cmtkparams, cmtk.mat2dof, cmtkparams2affmat