Make in-memory CMTK registration list from affine matrix or CMTK parameters

Make in-memory CMTK registration list from affine matrix or CMTK parameters

cmtkreglist(x, centre = c(0, 0, 0), reference = "dummy",
  floating = "dummy")

Arguments

x
5x3 matrix of CMTK registration parameters OR 4x4 homogeneous affine matrix

centre
Optional centre of rotation passed to affmat2cmtkparams when decomposing 4x4 affine matrix

reference, floating
Path to refererence and floating images.

Value

list of class cmtkreg containing registration parameters suitable for write.cmtkreg

Details

Note that this uses the modern CMTK notation of floating_study rather than model_study as used by IGSParamsToIGSRegistration (which results in an implicit inversion by CMTK tools).

Note that the reference and floating fields have no impact on the transformation encoded in the resultant .list folder and can be overridden on the command line of CMTK tools.

See also

write.cmtkreg, affmat2cmtkparams, cmtkreg