Get topological features per each node

Assigns to each node a distance from cell body.

get_topo_features(n)

get_distance_to_soma(n)

Arguments

n

neuron object with soma

Value

list with distance and Reversed Strahler order features per node. vector with distances from soma

See also

Examples

get_topo_features(Cell07PNs[[1]])
#> $distance
#>   [1]   0.000000   2.891293   5.703212   7.723529  10.314772  12.203266
#>   [7]  14.613359  16.836603  18.522564  21.691392  24.088312  26.019016
#>  [13]  28.017067  30.812738  33.730963  37.287632  40.035391  43.014518
#>  [19]  46.913639  48.974197  51.765066  54.348195  56.346939  58.879558
#>  [25]  60.813022  63.894453  67.438223  70.995234  74.843541  77.504566
#>  [31]  79.784778  81.510994  83.049232  84.198101  85.705318  87.548086
#>  [37]  89.142351  90.314442  91.953660  93.378772  95.232832  97.874426
#>  [43]  85.713709  87.253311  89.419181  90.432009  92.473193  94.035976
#>  [49]  95.606273  97.375491  98.535834  99.195186 100.232652 101.237106
#>  [55] 105.239372 106.165457 107.005080 109.377553 110.340120  99.136544
#>  [61] 100.090064 100.875547  96.617044  98.002667 100.116655 101.569852
#>  [67] 104.017738 105.073912 106.459371 108.090653 111.057838 113.756003
#>  [73] 114.938956 116.333316 118.383879 119.326207 120.333314 121.402024
#>  [79] 121.939041 122.872372 122.140038 123.044352 124.396046 125.413416
#>  [85] 125.856894 119.892363 121.440277 123.221099 125.000460 127.333104
#>  [91] 129.047171 130.896452 134.654689 136.962823 141.589070 143.105171
#>  [97] 144.353221 146.760090 148.923239 149.550540 152.579000 152.873903
#> [103] 151.712116 155.154894 157.770555 160.517514 161.676557 163.330378
#> [109] 165.404046 166.330489 167.712941 168.814914 167.301730 168.467964
#> [115] 169.947808 171.004423 171.890998 171.029425 172.974349 173.628966
#> [121] 174.499355 173.794180 174.784306 175.915157 176.982712 178.000973
#> [127] 179.231757 180.618743 181.983596 182.726752 183.777074 184.355191
#> [133] 185.177770 185.775106 164.484244 166.126191 167.551499 168.764320
#> [139] 169.766696 170.825411 172.238196 173.250560 175.399201 176.685745
#> [145] 177.958924 179.671852 181.392164 182.260234 177.508785 178.726600
#> [151] 179.660677 180.936156 182.307256 183.045772 166.310862 167.554380
#> [157] 169.307693 170.621393 171.870419 172.529211 173.672155 174.998924
#> [163] 176.168109 176.905346 177.395288 173.449288 174.459197 175.670626
#> [169] 176.264856 177.016410 177.777626 178.222027 177.010832 178.424873
#> [175] 179.662183 180.861082 182.080677 183.656900 184.527619 186.085904
#> 
#> $rso
#>   [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 2 2 2
#>  [38] 2 2 2 2 2 0 0 0 0 0 0 1 1 1 2 2 2 2 2 2 2 2 2 2 2 0 0 0 0 0 0 0 0 0 0 0 0
#>  [75] 0 1 1 1 2 2 2 2 2 2 2 0 0 0 0 0 0 0 0 0 0 2 0 0 1 2 2 2 0 0 0 0 0 0 1 2 2
#> [112] 2 1 1 1 2 2 1 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 0 1 1 1 1 1 1 1 1 2 2 2 2 2
#> [149] 2 2 2 2 2 2 1 1 1 1 1 1 2 2 2 2 2 1 1 1 1 2 2 2 2 2 2 2 2 2 2 2
#> 
get_distance_to_soma(Cell07PNs[[1]])
#>   [1]   0.000000   2.891293   5.703212   7.723529  10.314772  12.203266
#>   [7]  14.613359  16.836603  18.522564  21.691392  24.088312  26.019016
#>  [13]  28.017067  30.812738  33.730963  37.287632  40.035391  43.014518
#>  [19]  46.913639  48.974197  51.765066  54.348195  56.346939  58.879558
#>  [25]  60.813022  63.894453  67.438223  70.995234  74.843541  77.504566
#>  [31]  79.784778  81.510994  83.049232  84.198101  85.705318  87.548086
#>  [37]  89.142351  90.314442  91.953660  93.378772  95.232832  97.874426
#>  [43]  85.713709  87.253311  89.419181  90.432009  92.473193  94.035976
#>  [49]  95.606273  97.375491  98.535834  99.195186 100.232652 101.237106
#>  [55] 105.239372 106.165457 107.005080 109.377553 110.340120  99.136544
#>  [61] 100.090064 100.875547  96.617044  98.002667 100.116655 101.569852
#>  [67] 104.017738 105.073912 106.459371 108.090653 111.057838 113.756003
#>  [73] 114.938956 116.333316 118.383879 119.326207 120.333314 121.402024
#>  [79] 121.939041 122.872372 122.140038 123.044352 124.396046 125.413416
#>  [85] 125.856894 119.892363 121.440277 123.221099 125.000460 127.333104
#>  [91] 129.047171 130.896452 134.654689 136.962823 141.589070 143.105171
#>  [97] 144.353221 146.760090 148.923239 149.550540 152.579000 152.873903
#> [103] 151.712116 155.154894 157.770555 160.517514 161.676557 163.330378
#> [109] 165.404046 166.330489 167.712941 168.814914 167.301730 168.467964
#> [115] 169.947808 171.004423 171.890998 171.029425 172.974349 173.628966
#> [121] 174.499355 173.794180 174.784306 175.915157 176.982712 178.000973
#> [127] 179.231757 180.618743 181.983596 182.726752 183.777074 184.355191
#> [133] 185.177770 185.775106 164.484244 166.126191 167.551499 168.764320
#> [139] 169.766696 170.825411 172.238196 173.250560 175.399201 176.685745
#> [145] 177.958924 179.671852 181.392164 182.260234 177.508785 178.726600
#> [151] 179.660677 180.936156 182.307256 183.045772 166.310862 167.554380
#> [157] 169.307693 170.621393 171.870419 172.529211 173.672155 174.998924
#> [163] 176.168109 176.905346 177.395288 173.449288 174.459197 175.670626
#> [169] 176.264856 177.016410 177.777626 178.222027 177.010832 178.424873
#> [175] 179.662183 180.861082 182.080677 183.656900 184.527619 186.085904