Interconvert pixel and physical coordinates

xyzpos converts pixel coordinates to physical coordinates ijkpos converts physical coordinates to pixel coordinates

xyzpos(d, ijk)

ijkpos(d, xyz, roundToNearestPixel = TRUE)

Arguments

d
An im3d object defining a physical space

ijk
an Nx3 matrix of pixel coordinates (1-indexed)

xyz
Nx3 matrix of physical coordinates

roundToNearestPixel
Whether to round calculated pixel coordinates to nearest integer value (i.e. nearest pixel). default: TRUE

Value

Nx3 matrix of physica l or pixel coordinates

See also

ind2coord

Other im3d: as.im3d, boundingbox, im3d-io, im3d, imexpand.grid, imslice, is.im3d, mask, origin, projection, threshold, unmask, voxdims

Examples

# make an emty im3d
d=im3d(,dim=c(20,30,40),origin=c(10,20,30),voxdims=c(1,2,3))
# check round trip for origin
stopifnot(all.equal(ijkpos(d,xyzpos(d,c(1,1,1))), c(1,1,1)))