Check whether a file is in NeuroML format

This will check a file on disk to see if it is in NeuroML format. Some prechecks (optionally taking place on a supplied raw vector of bytes) should weed out nearly all true negatives and identify many true positives without having to read/parse the file header.

is.neuroml(f, bytes = NULL)

Arguments

f
path to a file on disk

bytes
optional raw vector of bytes used for prechecks