Check if a file is in the raw image format used by Hanchuan Peng's Vaa3D

See http://www.vaa3d.org/ https://svn.janelia.org/penglab/projects/vaa3d/trunk/imagej_io/v3draw_io_imagej/raw_reader.java

is.vaa3draw(f, bytes = NULL)

Arguments

f

A character vector specifying the path or a raw vector (see bytes).

bytes

optional raw vector of at least 24 bytes from the start of a single file (used in preference to reading file f).

Details

Note that multiple files can be checked when a character vector of length > 1 is provided, but only one file can be checked when a raw byte array is provided.