List of 20 Kenyon Cells from Chiang et al 2011 converted to dotprops objects

This R list (which has additional class neuronlist) contains 20 skeletonized Drosophila Kenyon cells as dotprops objects. Original data is due to Chiang et al. 2011, who have generously shared their raw data at http://flycircuit.tw. Image registration and further processing was carried out by Greg Jefferis.

References

[1] Chiang A.S., Lin C.Y., Chuang C.C., Chang H.M., Hsieh C.H., Yeh C.W., Shih C.T., Wu J.J., Wang G.T., Chen Y.C., Wu C.C., Chen G.Y., Ching Y.T., Lee P.C., Lin C.Y., Lin H.H., Wu C.C., Hsu H.W., Huang Y.A., Chen J.Y., et al. (2011). Three-dimensional reconstruction of brain-wide wiring networks in Drosophila at single-cell resolution. Curr Biol 21 (1), 1--11.

See also

head.neuronlist, with.neuronlist, plot3d.neuronlist, plot3d.dotprops, dotprops

Other nat-data: Cell07PNs, MBL.surf

Examples

head(kcs20)

#>                                       gene_name          Name  idid soma_side
#> FruMARCM-M001205_seg002 FruMARCM-M001205_seg002  fru-M-500112  1024         L
#> GadMARCM-F000122_seg001 GadMARCM-F000122_seg001 Gad1-F-900005 10616         L
#> GadMARCM-F000050_seg001 GadMARCM-F000050_seg001 Gad1-F-100010  8399         R
#> GadMARCM-F000142_seg002 GadMARCM-F000142_seg002 Gad1-F-300043 10647         L
#> FruMARCM-F000270_seg001 FruMARCM-F000270_seg001  fru-F-400045  9758         L
#> FruMARCM-F001115_seg002 FruMARCM-F001115_seg002  fru-F-300059  6182         R
#>                         flipped    Driver Gender        X         Y         Z
#> FruMARCM-M001205_seg002   FALSE  fru-Gal4      M 361.4849  95.04480  84.10259
#> GadMARCM-F000122_seg001   FALSE Gad1-Gal4      F 367.8332 105.86755  94.73446
#> GadMARCM-F000050_seg001    TRUE Gad1-Gal4      F 382.8279  61.73213  97.28057
#> GadMARCM-F000142_seg002   FALSE Gad1-Gal4      F 349.5917  78.18986  96.69280
#> FruMARCM-F000270_seg001   FALSE  fru-Gal4      F 387.5236 114.80344  87.84156
#> FruMARCM-F001115_seg002    TRUE  fru-Gal4      F 352.0216 121.72034 100.52308
#>                                        exemplar cluster  idx  type
#> FruMARCM-M001205_seg002 FruMARCM-M001205_seg002       9  156 gamma
#> GadMARCM-F000122_seg001 GadMARCM-F000122_seg001      70 1519 gamma
#> GadMARCM-F000050_seg001 GadMARCM-F000050_seg001      57 1132    ab
#> GadMARCM-F000142_seg002 GadMARCM-F000142_seg002      71 1535  apbp
#> FruMARCM-F000270_seg001 FruMARCM-F000270_seg001      64 1331    ab
#> FruMARCM-F001115_seg002 FruMARCM-F001115_seg002      44  795    ab

table(with(kcs20, type))

#> 
#>    ab  apbp gamma 
#>    10     3     7 

nopen3d()

#> glX 
#>   9 

# see plot3d.neuronlist documentation for more details
plot3d(kcs20, col=type)