Create a neuronlist from zero or more neurons

neuronlist objects consist of a list of neuron objects (usually of class neuron or dotprops) along with an optional attached dataframe containing information about the neurons. neuronlist objects can be indexed using their name or the number of the neuron like a regular list. Both the list itself and the attached data.frame must have the same unique (row)names. If the [ operator is used to index the list, the attached dataframe will also be subsetted.

It is perfectly acceptable not to pass any parameters, generating an empty neuronlist

neuronlist(..., DATAFRAME = NULL)

Arguments

...
objects to be turned into a list

DATAFRAME
an optional data.frame to attach to the neuronlist containing information about each neuron.

Value

A new neuronlist object.

See also

Examples

# generate an empty neuronlist
nl=neuronlist()
# slice an existing neuronlist with regular indexing
kcs5=kcs20[1:5]

# extract a single neuron from a neuronlist
n1=Cell07PNs[[1]]

# list all methods for neuronlist objects
methods(class='neuronlist')

#>  [1] -                  [                  [<-                *                 
#>  [5] /                  +                  as.data.frame      as.neuronlistfh   
#>  [9] c                  data.frame<-       dimnames           dotprops          
#> [13] droplevels         head               intersect          mirror            
#> [17] nvertices          plot               plot3d             potential_synapses
#> [21] prune              setdiff            subset             summary           
#> [25] tail               union              with               xform             
#> [29] xyzmatrix          xyzmatrix<-       
#> see '?methods' for accessing help and source code