Open customised rgl window

Open customised rgl window

nopen3d(bgcol = "white", FOV = 0, ...)

Arguments

bgcol

background colour

FOV

field of view

...

additional options passed to open3d

Value

current rgl device

Details

Pan with right button (Ctrl+click), zoom with middle (Alt/Meta+click) button. On a Mac trackpad, pan with two fingers left-right, zoom with two fingers in-out. Defaults to a white background and orthogonal projection (FOV=0)

Note that sometimes (parts of) objects seem to disappear after panning and zooming. See help for pan3d.

rgl and plotly have quite different models for how to handle the active plot. nopen3d and nclear3d allow you to treat them more similarly. Use them wherever you use the rgl clear3d and open3d commands and your could she be able to run with both plotly or rgl as the plotengine.

See also