Find the number of vertices in an object (or each element of a neuronlist)

Find the number of vertices in an object (or each element of a neuronlist)

nvertices(x, ...)

# S3 method for default
nvertices(x, ...)

# S3 method for neuronlist
nvertices(x, ...)

Arguments

x

An object with 3d vertices (e.g. neuron, surface etc)

...

Additional arguments passed to methods (currently ignored)

Value

an integer number of vertices (or a vector of length equal to a neuronlist)

Examples

nvertices(Cell07PNs[[1]])

#> [1] 180

nvertices(kcs20)

#> FruMARCM-M001205_seg002 GadMARCM-F000122_seg001 GadMARCM-F000050_seg001 
#>                     284                     297                     374 
#> GadMARCM-F000142_seg002 FruMARCM-F000270_seg001 FruMARCM-F001115_seg002 
#>                     355                     369                     401 
#> FruMARCM-M001051_seg002 GadMARCM-F000423_seg001 ChaMARCM-F000586_seg002 
#>                     413                     415                     418 
#> FruMARCM-M001339_seg001 GadMARCM-F000476_seg001 FruMARCM-F000085_seg001 
#>                     436                     453                     457 
#> FruMARCM-F000706_seg001 FruMARCM-M000842_seg002 FruMARCM-F001494_seg002 
#>                     459                     457                     481 
#> FruMARCM-F000188_seg001 GadMARCM-F000071_seg001 FruMARCM-M000115_seg001 
#>                     484                     495                     477 
#> GadMARCM-F000442_seg002 FruMARCM-F001929_seg001 
#>                     506                     490