plot3d methods for different nat objects

These methods enable nat objects including neuronlists and dotprops objects to be plotted in 3D. See the help for each individual method for details along with the help for the generic in the rgl package.

See also

Examples

# all known plot3d methods
methods("plot3d")
#>  [1] plot3d.ashape3d*      plot3d.boundingbox*   plot3d.character*    
#>  [4] plot3d.cmtkreg*       plot3d.default*       plot3d.deldir*       
#>  [7] plot3d.dotprops*      plot3d.formula*       plot3d.function*     
#> [10] plot3d.hxsurf*        plot3d.lm*            plot3d.mesh3d*       
#> [13] plot3d.neuron*        plot3d.neuronlist*    plot3d.ngraph*       
#> [16] plot3d.rglWebGL*      plot3d.rglbackground* plot3d.rglbboxdeco*  
#> [19] plot3d.rglobject*     plot3d.rglscene*      plot3d.rglsubscene*  
#> [22] plot3d.templatebrain* plot3d.tri*           plot3d.triSht*       
#> see '?methods' for accessing help and source code

# up to date list of all plot3d nethods in this package
intersect(methods("plot3d"), ls(asNamespace("nat")))
#> [1] "plot3d.boundingbox" "plot3d.character"   "plot3d.cmtkreg"    
#> [4] "plot3d.dotprops"    "plot3d.hxsurf"      "plot3d.neuron"     
#> [7] "plot3d.neuronlist"  "plot3d.ngraph"