Plot the domain of a CMTK registration

Plot the domain of a CMTK registration

# S3 method for cmtkreg
plot3d(x, ..., gridlines = FALSE, plotengine = getOption("nat.plotengine"))

Arguments

x

A cmtk registration (the path to the registration folder on disk) or the resulting of reading one in with read.cmtkreg.

...

Additional arguments passed to plot3d

gridlines

Whether to display gridlines when using plotly as the backend plotting engine (default: FALSE)

plotengine

the plotting backend engine to use either 'rgl' or 'plotly'.

See also

Examples

# \donttest{
testdatadir=system.file("tests/testthat/testdata/cmtk", package="nat")
regpath=file.path(testdatadir,'FCWB_JFRC2_01_warp_level-01.list/')
# only run this if file is present (not always installed)
if(file.exists(regpath)){
plot3d(cmtkreg(regpath))

# or read registration into memory if you want to work with it
reg=read.cmtkreg(regpath)
plot3d(reg)
}

# }