Find which points of an object are inside a surface

Find which points of an object are inside a surface

pointsinside(x, surf, ...)

# S3 method for default
pointsinside(x, surf, ..., rval = c("logical", "distance",
  "mesh3d"))

Arguments

x
an object with 3D points.

surf
The reference surface - either a mesh3d object or any object that can be converted using as.mesh3d including hxsurf and ashape3d objects.

...
additional arguments for methods, eventually passed to as.mesh3d.

rval
what to return.

Value

A vector of logical values or distances (positive inside, negative outside) equal to the number of points in x or the mesh3d object returned by Rvcg::vcgClost.

Details

Note that hxsurf surface objects will be converted to mesh3d before being passed to Rvcg::vcgClost, so if you are testing repeatedly against the same surface, it may make sense to pre-convert.

Note also that if the point is some distance (> 2 twice the diagonal boundingbox of the mesh) then the distance will be returned as NA (or 1e12 for older versions of the Rvcg package). This behaviour is defined by the Rvcg::vcgClost function.

Examples

# check if the vertices in these neurons are inside the mushroom body
# surface object
inout=pointsinside(kcs20, surf=MBL.surf)
table(inout)

#> inout
#> FALSE  TRUE 
#>  1869  6652 

# be a bit more lenient and include points less than 5 microns from surface
inout5=pointsinside(kcs20, surf=MBL.surf, rval='distance') > -5
table(inout5)

#> inout5
#> FALSE  TRUE 
#>   236  8285 

# show which points are in or out
points3d(xyzmatrix(kcs20), col=ifelse(inout5, 'red', 'black'))
plot3d(MBL.surf, alpha=.3)