Synchronise a remote object

Synchronise a remote object

remotesync(x, remote = attr(x, "remote"), download.missing = TRUE,
  delete.extra = FALSE, ...)

# S3 method for neuronlistfh
remotesync(x, remote = attr(x, "remote"),
  download.missing = FALSE, delete.extra = FALSE, indices = NULL,
  update.object = TRUE, ...)

Arguments

x
Object to synchronise with a remote URL

remote
The remote URL to update from

download.missing
Whether to download missing objects (default TRUE)

delete.extra
Whether to delete objects (default TRUE)

…
Additional arguments passed to methods

indices
Character vector naming neurons to update (default indices=NULL implies all neurons).

update.object
Whether to update the neuronlistfh object itself on disk (default TRUE). Note that this assumes that the neuronlistfh object has not been renamed after it was downloaded.

Value

The updated neuronlistfh object (invisibly)

See also

Other neuronlistfh: [.neuronlistfh, neuronlistfh, read.neuronlistfh, write.neuronlistfh

Examples

not_run({
  kcs20=read.neuronlistfh('http://flybrain.mrc-lmb.cam.ac.uk/si/nblast/flycircuit/kcs20.rds')
  # update object from the web
  kcs20=remotesync(kcs20)
  # download all neurons with significant innervation of the vertical lobe
  mbvl_neurons=subset(kcs20, (MB_VL_R+MB_VL_L)>200, rval='names')
  kcs20=remotesync(kcs20, indices=mbvl_neurons, download.missing=TRUE)
})