Simplify a registration list

Simplify a registration list

simplify_reglist(reg, as.cmtk = NULL)

Arguments

reg
A registration list (reglist) containing one or more transformations.

as.cmtk
Whether to convert to a vector of CMTK format registrations (see cmtkreg). The default value of as.cmtk=NULL converts all registrations to CMTK if any one registration is in CMTK format (thus enabling them to be applied by CMTK tools in a single call). See details.

Details

This function

inverts any affine matrices with attribute "swap"

collapses multiple affine matrices into a single affine

optionally converts all registrations to CMTK on disk registrations when possible.

Note that if any of the registrations are in CMTK format, the default behaviour is to try to convert all of the other registrations into CMTK format to enable them to be passed to CMTK in a single command. If as.cmtk=TRUE then there will be an error if this is not possible.

See also

reglist, xform, cmtkreg