Find the Strahler order of each point in a neuron

The Strahler order will be 1 for each tip segment and then 1 + the maximum of the Strahler order of each parent segment for internal segments. Branch points will have the Strahler order of the closest segment to the root of which they are part.

strahler_order(x)

Arguments

x
A neuron

Value

A list containing

points Vector of integer Strahler orders for each point in the neuron

segments Vector of integer Strahler orders for each segment in the neuron

Details

It is vital that the root of the neuron is valid since this determines the flow direction for calculation of the Strahler order. At present the function is not defined for neurons with multiple subtrees.

Internally, this function uses segmentgraph to find a reduced segmentgraph for the neuron.

References

https://en.wikipedia.org/wiki/Strahler_number

See also