Write out a neuronlistfh object to an RDS file

Write out a neuronlistfh object to an RDS file

write.neuronlistfh(x, file = attr(x, "file"), overwrite = FALSE, ...)

Arguments

x

The neuronlistfh object to write out

file

Path where the file will be written (see details)

overwrite

Whether to overwrite an existing file

&#8230;

Additional paramaters passed to saveRDS

Details

This function writes the main neuronlistfh object to disk, but makes no attempt to touch/verify the associated object files.

if file is not specified, then the function will first check if x has a 'file' attribute. If that does not exist, then attr(x,'db')@dir, the backing filehash database directory, is inspected. The save path file will then be constructed by taking the directory one up from the database directory and using the name of the neuronlistfh object with the suffix '.rds'. e.g. write.neuronlistfh(kcs20) with db directory '/my/path/dps/data' will be saved as '/my/path/dps/kcs20.rds'

Note that if x has a 'file' attribute (set by read.neuronlistfh) then this will be removed before the file is saved (since the file attribute must be set on read to ensure that we know exactly which file on disk was the source of the object in memory).

See also

saveRDS

Other neuronlistfh: [.neuronlistfh, neuronlistfh, read.neuronlistfh, remotesync