Cell07PNs: 40 Sample Projection Neurons from Jefferis, Potter et al 2007

These R lists (which have additional class neuronlist) contain 40 traced olfactory projection neurons from Jefferis, Potter et al 2007 that have been transformed onto the IS2 template brain (Cachero, Ostrovsky et al 2010).

References

Jefferis G.S.X.E., Potter C.J., Chan A.M., Marin E.C., Rohlfing T., Maurer C.R.J., and Luo L. (2007). Comprehensive maps of Drosophila higher olfactory centers: spatially segregated fruit and pheromone representation. Cell 128 (6), 1187--1203. doi:10.1016/j.cell.2007.01.040

Cachero S., Ostrovsky A.D., Yu J.Y., Dickson B.J., and Jefferis G.S.X.E. (2010). Sexual dimorphism in the fly brain. Curr Biol 20 (18), 1589--601. doi:10.1016/j.cub.2010.07.045

See also

Examples

head(Cell07PNs)

#>         Brain RegistrationScore Notes Glomerulus Batch Directory Traced
#> EBH11R EBH11R                 4              DA1   PN2    unsure    yes
#> EBH20L EBH20L                 4              DL3   PN2   unsure2    yes
#> EBH20R EBH20R                 4              DA1   PN2   unsure2    yes
#> EBI12L EBI12L                 4              DA1   PN2   unsure3    yes
#> EBI22R EBI22R                 4              DL3   PN2   unsure3    yes
#> EBJ23L EBJ23L                 4              DL3   PN2   unsure4    yes
#>        Scored.By Sex Include     ID HaveWarp HaveAsc
#> EBH11R       ACH   M         EBH11R     TRUE    TRUE
#> EBH20L       ACH             EBH20L     TRUE    TRUE
#> EBH20R       ACH   M         EBH20R     TRUE    TRUE
#> EBI12L       ACH   F         EBI12L     TRUE    TRUE
#> EBI22R       ACH             EBI22R     TRUE    TRUE
#> EBJ23L       ACH             EBJ23L     TRUE    TRUE
#>                                                 TraceFile AscBatch Status
#> EBH11R /GD/projects/PN2/TransformedTraces/DA1/EBH11R.tasc      New       
#> EBH20L /GD/projects/PN2/TransformedTraces/DL3/EBH20L.tasc      New       
#> EBH20R /GD/projects/PN2/TransformedTraces/DA1/EBH20R.tasc      New       
#> EBI12L /GD/projects/PN2/TransformedTraces/DA1/EBI12L.tasc      New       
#> EBI22R /GD/projects/PN2/TransformedTraces/DL3/EBI22R.tasc      New       
#> EBJ23L /GD/projects/PN2/TransformedTraces/DL3/EBJ23L.tasc      New       
#>        GlomSeq NumNAs MBP1 MBP2 LHBP PNType Seq nTrees StartPoint
#> EBH11R       1      0   34   48   75    iPN  13      1          1
#> EBH20L       1      0   29   29   77    iPN  15      1          1
#> EBH20R       2      0   24   80  147    iPN  16      1          1
#> EBI12L       3      0   29   59   89    iPN  18      1          1
#> EBI22R       2      0   34   34   56    iPN  20      1          1
#> EBJ23L       3      0   23   23   52    iPN  21      1          1
#>                  CreatedAt
#> EBH11R 2006-01-17 15:21:14
#> EBH20L 2006-01-17 15:21:14
#> EBH20R 2006-01-17 15:21:14
#> EBI12L 2006-01-17 15:21:15
#> EBI22R 2006-01-17 15:21:15
#> EBJ23L 2006-01-17 15:21:15
#>                                                                                             WarpFile
#> EBH11R  /GD/projects/PN2/allreg/warp/unsure/average-goodbrains_EBH11R101_warp_m0g40c4e1e-1x16r3.list
#> EBH20L /GD/projects/PN2/allreg/warp/unsure2/average-goodbrains_EBH20L101_warp_m0g40c4e1e-1x16r3.list
#> EBH20R /GD/projects/PN2/allreg/warp/unsure2/average-goodbrains_EBH20R101_warp_m0g40c4e1e-1x16r3.list
#> EBI12L /GD/projects/PN2/allreg/warp/unsure3/average-goodbrains_EBI12L101_warp_m0g40c4e1e-1x16r3.list
#> EBI22R /GD/projects/PN2/allreg/warp/unsure3/average-goodbrains_EBI22R101_warp_m0g40c4e1e-1x16r3.list
#> EBJ23L /GD/projects/PN2/allreg/warp/unsure4/average-goodbrains_EBJ23L101_warp_m0g40c4e1e-1x16r3.list

table(with(Cell07PNs,Glomerulus))

#> 
#>  DA1  DL3 DP1m VA1d 
#>   11   10    8   11