Convert an object to an rgl mesh3d

as.mesh3d.ashape3d converts an alphashape3d::ashape3d object into a nat/rgl compatible mesh3d surface

Note that this provides a link to the Rvcg package

# S3 method for ashape3d
as.mesh3d(x, ...)

# S3 method for hxsurf
as.mesh3d(x, Regions = NULL, material = NULL,
  drop = TRUE, ...)

Arguments

x
Object to convert to mesh3d

...
Additional arguments for methods

Regions
Character vector or regions to select from hxsurf object

material
rgl materials such as color

drop
Whether to drop unused vertices (default TRUE)

Value

a mesh3d object which can be plotted and manipulated using rgl and nat packages.

Details

An alpha shape is a generalisation of a convex hull enclosing a set of points. Unlike a convex hull, the resultant surface can be partly concave allowing the surface to more closely follow the set of points. The parameter alpha

See also

ashape3d, mesh3d as.mesh3d, tmesh3d, as.hxsurf, read.hxsurf

Other hxsurf: as.hxsurf, materials, plot3d.hxsurf, read.hxsurf, subset.hxsurf, write.hxsurf

Examples

library(alphashape3d)
kcs20.a=ashape3d(xyzmatrix(kcs20), alpha = 10)
plot(kcs20.a)

#> Device  30  : alpha =  10 

# convert to mesh3d
kcs20.mesh=as.mesh3d(kcs20.a)
clear3d()
wire3d(kcs20.mesh)
plot3d(kcs20, col=type, lwd=2)