dotprops: Neurons as point clouds with tangent vectors (but no connectivity)

dotprops: Neurons as point clouds with tangent vectors (but no connectivity)

dotprops makes dotprops representation from raw 3D points (extracting vertices from S3 objects that have them)

dotprops.dotprops will default to the original vale of k and copy over all attributes that are not set by dotprops.default.

dotprops.neuronlist will run for every object in the neuronlist using nlapply. ... arguments will be passed to nlapply in addition to the named argument OmitFailures.

is.dotprops(x)

as.dotprops(x, ...)

dotprops(x, ...)

# S3 method for character
dotprops(x, ...)

# S3 method for dotprops
dotprops(x, k = attr(x, "k"), ...)

# S3 method for im3d
dotprops(x, ...)

# S3 method for neuronlist
dotprops(x, ..., OmitFailures = NA)

# S3 method for neuron
dotprops(x, Labels = NULL, resample = NA, ...)

# S3 method for default
dotprops(x, k = NULL, Labels = NULL, na.rm = FALSE, ...)

Arguments

x

Object to be tested/converted

...

Additional arguments passed to methods

k

Number of nearest neighbours to use for tangent vector calculation (set to k=20 when passed NULL)

OmitFailures

Whether to omit neurons for which FUN gives an error. The default value (NA) will result in nlapply stopping with an error message the moment there is an eror. For other values, see details.

Labels

Vector of labels for each point or NULL to accept class-specific default behaviour for different S3 classes, TRUE always to use labels when incoming object has them and FALSE never to use labels.

resample

When finite, a new length to which all segmented edges will be resampled. See resample.neuron.

na.rm

Whether to remove NA points (default FALSE)

Details

k will default to 20 nearest neighbours when unset (i.e. when it has default value of NA) unless x is a dotprops object (when the original value of k is reused).

References

The dotprops format is essentially identical to that developed in:

Masse N.Y., Cachero S., Ostrovsky A., and Jefferis G.S.X.E. (2012). A mutual information approach to automate identification of neuronal clusters in Drosophila brain images. Frontiers in Neuroinformatics 6 (00021). http://dx.doi.org/10.3389/fninf.2012.00021

See also

nlapply