dotprops: Neurons as point clouds with tangent vectors (but no connectivity)

dotprops makes dotprops representation from raw 3D points (extracting vertices from S3 objects that have them)

dotprops.character makes dotprops objects from one or more files on disk (typically binary segmentations saved as NRRDs). x can a vector of paths or be a directory (in which case pattern can be used to restrict the files to read). The ... argument is passed first to nlapply (if there is more than one file) and then to dotprops.default.

dotprops.dotprops will default to the original vale of k and copy over all attributes that are not set by dotprops.default.

dotprops.neuronlist will run for every object in the neuronlist using nlapply. ... arguments will be passed to nlapply in addition to the named argument OmitFailures.

is.dotprops(x)

as.dotprops(x, ...)

dotprops(x, ...)

# S3 method for character
dotprops(x, pattern = NULL, OmitFailures = NA, ...)

# S3 method for dotprops
dotprops(x, k = attr(x, "k"), ...)

# S3 method for im3d
dotprops(x, ...)

# S3 method for neuronlist
dotprops(x, ..., OmitFailures = NA)

# S3 method for neuron
dotprops(x, Labels = NULL, resample = NA, topo = FALSE, ...)

# S3 method for default
dotprops(x, k = NULL, Labels = NULL, na.rm = FALSE, topo_features = NULL, ...)

Arguments

x
Object to be tested/converted

...
Additional arguments passed to methods

pattern
an optional regular expression. Only file names which match the regular expression will be returned.

OmitFailures
Whether to omit neurons for which FUN gives an error. The default value (NA) will result in nlapply stopping with an error message the moment there is an error. For other values, see details.

k
Number of nearest neighbours to use for tangent vector calculation (set to k=20 when passed NULL)

Labels
Vector of labels for each point e.g. identifying axon vs dendrite. The default value NULL will produce class-specific default behaviour for different classes of input object, TRUE always uses labels when an incoming object has them and FALSE never uses labels.

resample
When finite, a new length to which all segmented edges will be resampled. See resample.neuron.

topo
flag that says whether or not to add topological features (inverted Strahler's Order and distance from soma)

na.rm
Whether to remove NA points (default FALSE)

topo_features
topological features of each dotprop

Details

k will default to 20 nearest neighbours when unset (i.e. when it has default value of NA) unless x is a dotprops object (when the original value of k is reused).

References

The dotprops format is essentially identical to that developed in:

Masse N.Y., Cachero S., Ostrovsky A., and Jefferis G.S.X.E. (2012). A mutual information approach to automate identification of neuronal clusters in Drosophila brain images. Frontiers in Neuroinformatics 6 (00021). doi: 10.3389/fninf.2012.00021

See also

Examples

if (FALSE) {
# process a single file on disk
dp=dotprops.character('~/skeleton-nrrds/file01.nrrd', k=5)
# process a whole directory of files
dps=dotprops.character('~/skeleton-nrrds/', OmitFailures=T, k=5)
}