Remove plotted neurons or other 3D objects

The normal usage will not specify x in which case the last neurons plotted by plot3d.neuronlist or any of its friends will be removed.

npop3d(x, slow = FALSE, type = "shapes")

Arguments

x
rgl ids of objects to remove

slow
Whether to remove neurons one by one (slowly) default: FALSE

type
Type of objects to remove see pop3d.

See also