3D plots of dotprops objects using rgl package

3D plots of dotprops objects using rgl package

# S3 method for dotprops
plot3d(x, scalevecs = 1, alpharange = NULL,
  color = "black", PlotPoints = FALSE, PlotVectors = TRUE,
  UseAlpha = FALSE, ...)

Arguments

x

A dotprops object

scalevecs

Factor by which to scale unit vectors (numeric, default: 1.0)

alpharange

Restrict plotting to points with alpha values in this range to plot (default: null => all points). See dotprops for definition of alpha.

color

Character or numeric vector specifying colours for points/vectors. See details.

PlotPoints, PlotVectors

Whether to plot points and/or tangent vectors (logical, default: tangent vectors only)

UseAlpha

Whether to scale tangent vector length by the value of alpha

...

Additional arguments passed to points3d and/or segments3d

Value

invisible list of results of rgl plotting commands

Details

Tangent vectors are plotted by segments3d and centered on the relevant point. Points are plotted by points3d.

color will be recycled by points3d and segments3d. However in the special case that color has length equal to the number of points in x, then it will be duplicated before being passed to segments3d so that the result is that each vector is coloured uniformly according to color (since segments3d expects 2 colours for each line segment, blending them if they are different).

See also

dotprops, plot3d, points3d, segments3d

Examples