Summary statistics for neurons (e.g. cable length, number of nodes)

summary.neuronlist computes tree statistics for all the neurons in a neuronlist object

summary.neuron computes statistics for individual neurons

summary.dotprops computes statistics for individual neurons in dotprops format. Note the veclength argument.

# S3 method for neuronlist
summary(object, ...,
  include.attached.dataframe = FALSE)

# S3 method for neuron
summary(object, ...)

# S3 method for dotprops
summary(object, veclength = 1, ...)

Arguments

object
The neuron or neuronlist to summarise

...
For summary.neuronlist additional arguments passed on to summary methods for individual neurons

include.attached.dataframe
Whether to include the neuronlists attached metadata in the returned data.frame.

veclength
The vector length to assume for each segment so that a cable length estimate can be made.

Value

A data.frame summarising the tree properties of the neuron with columns

root

nodes

segments

branchpoints

endpoints

cable.length

See also

Examples

# summary for a whole neuronlist
summary(Cell07PNs)

#>        root nodes segments branchpoints endpoints cable.length
#> EBH11R    1   180       33           16        18     297.1763
#> EBH20L    1   200       26           12        15     327.0929
#> EBH20R    1   199       25           12        14     347.6153
#> EBI12L    1   169       23           11        13     294.4680
#> EBI22R    1   160       27           13        15     303.0150
#> EBJ23L    1   156       28           13        16     292.3298
#> EBJ3R     1   118       32           15        18     286.0233
#> EBN19L    1   162       31           15        17     314.7040
#> EBO15L    1   167       39           19        21     350.7743
#> EBO53L    1   175       26           12        15     314.9850
#> ECA34L    1   446      149           72        78     910.0081
#> ECB3L     1   366      124           57        68     936.4803
#> LI23L     1   188       29           14        16     236.5884
#> LIC2R     1   279       27           13        15     416.1551
#> LJ5L      1   154       29           14        16     241.9808
#> MC3B      1   118       27           13        15     280.7168
#> MH16L     1   171       22           10        13     261.7648
#> MM14L     1   216       23           11        13     305.3770
#> NA7L      1   121       13            6         8     186.6893
#> NH15L     1    83       26           12        15     212.3443
#> NH29B     1   101       34           16        19     231.5922
#> NI16L     1   122       28           13        16     226.2681
#> NIA8L     1   961       32           15        18     387.6834
#> NIA8R     1   797       25           12        14     332.4043
#> NNA9L     2  2481      171           84        88     991.7564
#> NNC4R     2  1902      125           61        65     864.2939
#> NNE1L     1  2500      163           78        86    1013.6281
#> OFD2L     1  2439      161           77        85     992.7070
#> OKC9R     1  2556      144           67        78    1013.9386
#> SDD8L     2  2452      151           74        78    1008.0937
#> SH21L     1   148       19            9        11     234.8228
#> SL20L     1   213       30           14        17     258.2600
#> TKC8R     2   605       11            4         8     254.0427
#> TL4R      1   135       25           11        15     211.1676
#> TS7L      1   176       33           16        18     244.8527
#> TT27R     1   168       30           14        17     226.0329
#> VA15R     1   147       17            8        10     213.8872
#> VA20R     1   160       21           10        12     215.0711
#> VB37L     1   162       14            6         9     218.7555
#> VB58L     1   154       17            8        10     231.9512

# including the attached data.frame with additional metadata
head(summary(Cell07PNs, include.attached.dataframe = FALSE))

#>        root nodes segments branchpoints endpoints cable.length
#> EBH11R    1   180       33           16        18     297.1763
#> EBH20L    1   200       26           12        15     327.0929
#> EBH20R    1   199       25           12        14     347.6153
#> EBI12L    1   169       23           11        13     294.4680
#> EBI22R    1   160       27           13        15     303.0150
#> EBJ23L    1   156       28           13        16     292.3298

# for a single regular format neuron
summary(Cell07PNs[[1]])

#>   root nodes segments branchpoints endpoints cable.length
#> 1    1   180       33           16        18     297.1763

# for a single dotprops format neuron
summary(kcs20[[1]])

#>   nodes cable.length
#> 1   284          284

# specify a different estimate for the cable length associated with a single
# point in the neuron
summary(kcs20[[1]], veclength=1.2)

#>   nodes cable.length
#> 1   284        340.8