Return voxel dimensions of an object

This would properly be thought of as the voxel spacing when voxels are assumed not to have a physical extent (only a location).

voxdims(x, ...)

# S3 method for im3d
voxdims(x, ...)

# S3 method for character
voxdims(x, ...)

# S3 method for default
voxdims(x, dims, ...)

Arguments

x
An im3d object with associated voxel dimensions, a path to or a 2 x 3 BoundingBox matrix.

...
Additional arguments for methods

dims
The number of voxels in each dimension when x is a BoundingBox matrix.

Value

A numeric vector of length 3, NA when missing.

Details

We follow Amira's convention of returning a voxel dimension equal to the bounding box size (rather than 0) for any dimension with only 1 voxel.

See also

boundingbox

Other im3d: as.im3d, boundingbox, im3d-coords, im3d-io, im3d, imexpand.grid, imslice, is.im3d, mask, origin, projection, threshold, unmask